+++ puppetdb/v3/__init__.py
@@ -20,3 +20,4 @@ import events
 import event_counts
 import aggregate_event_counts
 import server_time
+import version
